finwise_ai_mvp/backend/requirements.txt
fastapi
uvicorn
requests
transformers
torch
python-dotenv
yfinance